Cleaning
git-svn-id: file:///home/svn/shoola-omero/trunk@6036 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/imviewer/rnd/GraphicsPane.java
@@ -232,8 +232,8 @@ class GraphicsPane
     	 int width = codomainSlider.getPreferredSize().width;
     	 double size[][] =
          {{width, knobWidth/2, TableLayout.FILL, knobWidth/2},  // Columns
-          {knobHeight/2, TableLayout.FILL, knobHeight/2, knobHeight+2, 
-        	 TableLayout.PREFERRED, 5}}; // Rows
+          {knobHeight/2, TableLayout.FILL, (double) knobHeight/2, 
+        	 (double) knobHeight+2, TableLayout.PREFERRED, 5}}; // Rows
     	 p.setLayout(new TableLayout(size));
     	 p.add(codomainSlider, "0, 0, 0, 2");
     	 p.add(uiDelegate, "2, 1");
